Refactored tests into specific unit and integration testing suites
