+++ __tests__/client/integration/.gitkeep
@@ -1,0 +1,1 @@
+Future integration tests
